include "asterisk/zapata.h" instead of testing for the location of the header files.
On passing, add a cast to insure -Werror clean compilation
on FreeBSD 6.x, where time_t does not match %ld



git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@51293 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ apps/app_meetme.c
@@ -38,8 +38,8 @@ ASTERISK_FILE_VERSION(__FILE__, "$Revision$")
 #include <string.h>
 #include <unistd.h>
 #include <errno.h>
-#include <sys/ioctl.h>
-#include <zaptel/zaptel.h>
+
+#include "asterisk/zapata.h"
 
 #include "asterisk/lock.h"
 #include "asterisk/file.h"
@@ -1933,7 +1933,7 @@ bailoutandtrynormal:
 				      user->user_no,
 				      S_OR(user->chan->cid.cid_num, "<unknown>"),
 				      S_OR(user->chan->cid.cid_name, "<unknown>"),
-				      (now - user->jointime));
+				      (long)(now - user->jointime));
 		}
 
 		conf->users--;
